borg check: fixup for "deleting orphaned objs" msgs, fixes #3795
only output msgs if there is actually something to delete.
be more precise, show count of orphaned / superseded objects.

(cherry picked from commit d671e9acf2c0ace6fb692b6c517db1db19643043)

+++ src/borg/archive.py
@@ -1721,11 +1721,12 @@ class ArchiveChecker:
             if orphaned:
                 logger.error('{} orphaned objects found!'.format(len(orphaned)))
                 self.error_found = True
-            if self.repair:
-                logger.info('Deleting orphaned objects...')
+            if self.repair and unused:
+                logger.info('Deleting %d orphaned and %d superseded objects...' % (
+                    len(orphaned), len(self.possibly_superseded)))
                 for id_ in unused:
                     self.repository.delete(id_)
-                logger.info('Finished deleting orphaned objects.')
+                logger.info('Finished deleting orphaned/superseded objects.')
         else:
             logger.info('Orphaned objects check skipped (needs all archives checked).')
